Reorganización de las carpetas del repositorio. Agregar plugin rest spread.
