+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["es2015", "stage-2"]
+  "presets": ["es2015", "stage-2"],
+  "plugins": ["transform-object-rest-spread"]
 }
