+++ index.html
@@ -8,16 +8,16 @@
     <meta name="author" content="UTNianos" />
     <meta name="description" content="UTNianos 2.0 frontend" />
     <meta name="robots" content="index, follow" />
-    
+
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    
+
     <title>UTNianos</title>
-    
+
   </head>
-  
-  <body>    
+
+  <body>
     <div id="app"></div>
-    
+
   </body>
-  
+
 </html>
