+++ src/App/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Footer from './Footer';
+import Navbar from '../Navbar/Container';
+
+const App = ({children}) => (
+  <div>
+	  <Navbar />
+	  {children}
+	  <Footer />
+	</div>
+);
+
+export default App;
