+++ src/App/Footer.scss
