+++ src/App/Home.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () => (
+  <div>
+    <h1>Bienvenido a la nueva plataforma de UTNianos</h1>
+  </div>
+);
+
+export default Home;
