+++ src/Navbar/Navbar.js
@@ -1,0 +1,12 @@
+import React, { PropTypes } from 'react'
+
+const Navbar = () => {
+
+  return (
+	<ul>
+	  <li><a className="active" href="#">UTNianos</a></li>
+	</ul>
+  )
+}
+
+export default Navbar
