--- src/components/App.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import Footer from './Footer';
-import NavbarUTNianos from '../containers/NavbarUTNianos';
-
-const App = ({children}) => (    	  
-     <div>	
-		  <NavbarUTNianos />
-		  {children}
-		  <Footer />
-	 </div>  
-);
-
-export default App;
